Added src, errors, data, users, and admins directories. Organized directories per best practices.
