+++ backend/src/app.js
@@ -4,8 +4,8 @@ var path = require('path')
 var cookieParser = require('cookie-parser')
 var logger = require('morgan')
 
-var indexRouter = require('./routes/index')
-var usersRouter = require('./routes/users')
+var indexRouter = require('../routes/index')
+var usersRouter = require('../routes/users')
 
 var app = express()
 const { Pool } = require('pg')
