+++ backend/src/db/connection.js
@@ -1,5 +1,5 @@
 const environment = process.env.NODE_ENV || "development";
-const config = require("../knexfile")[environment];
+const config = require("../../knexfile")[environment];
 const knex = require("knex")(config);
 
 module.exports = knex;
